requirements.txt
# PlaudAI Uploader - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML - Gemini for Clinical Synopsis Generation
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0